Add generate content using video

+++ Gemini/src/utils/googleFileManager.js
@@ -1,10 +1,12 @@
 import { fileManager } from "../index.js";
+import { FileState } from "@google/generative-ai/server";
 import fs from "fs";
 
 // Upload to Google Files 
 const uploadToGoogleFiles = async (localFilePath, mimeType, displayName) => {
     try {
-        const uploadResults = await fileManager.uploadFile(
+        let uploadResults;
+        uploadResults = await fileManager.uploadFile(
             localFilePath,
             {
                 mimeType,
@@ -12,6 +14,18 @@ const uploadToGoogleFiles = async (localFilePath, mimeType, displayName) => {
             },
         );
         console.log("Upload Result: ", uploadResults);
+
+        // Wait till file state is ACTIVE 
+        while(uploadResults.file.state == FileState.PROCESSING) {
+            await new Promise(resolve => setTimeout(resolve, 5000));
+            uploadResults.file = await fileManager.getFile(uploadResults.file.name);
+            // console.log("Upload Result: ", uploadResults);
+        }
+
+        if (uploadResults.state !== "ACTIVE") {
+            throw new Error("File upload failed");
+        }
+
         fs.unlinkSync(localFilePath);
         return uploadResults;
